refactor(models): keep cli files within size limits
