+++ pkg/services/models/transports/cli/root_invoke.go
@@ -62,21 +62,6 @@ func (service *rootService) Invoke(cfg InvokeConfig) error {
 	return service.invokeInScope(cfg, scope.Scope, modelName, operation, text)
 }
 
-func inferGenericCLIModelOperation(modelName string) string {
-	switch strings.ToLower(strings.TrimSpace(modelName)) {
-	case strings.ToLower(modelinference.BuiltInModelNameLLM):
-		return modelinference.OperationOMNI
-	case strings.ToLower(modelinference.BuiltInModelNameASR):
-		return modelinference.OperationASR
-	case strings.ToLower(modelinference.BuiltInModelNameTTS):
-		return modelinference.OperationTTS
-	case strings.ToLower(modelinference.BuiltInModelNameEmbed):
-		return modelinference.OperationEMBED
-	default:
-		return ""
-	}
-}
-
 func (service *rootService) invokeInScope(
 	cfg InvokeConfig,
 	scope modelinference.RuntimeScopeRef,
